API Work - Spotify and Last FM
Rate limit for each of these is once/second. Feel free to take this code and fool around with polling the API, but make sure to do it no more frequently than once per second for each API. That is the rate limit for each of them. These python files have time.sleep(1) located in them for that purpose.
Don't take these files too seriously, but they're a starting point.