This gives clean, workable data
For the final csv that comes from this, all these columns can go straight to a database. Basically everything is already flattened, including the key attributes of artists and name (song name, that is). So the only things that are still nested but don't really need to be unwound are album_images and album_artists. Those can just be TEXT in postgres.
If someone can get this into postgres, that'd be huge.